Removes Backend class and moves logic and free functions to Launch file, updates tests
