+++ tests/test_util.py
@@ -7,9 +7,9 @@ from pvi.device import SignalR
 from pydantic import ValidationError
 
 from fastcs.attributes import AttrR, AttrRW
-from fastcs.backend import Backend
 from fastcs.controller import Controller
 from fastcs.datatypes import Bool, Enum, Float, Int, String
+from fastcs.launch import FastCS
 from fastcs.util import numpy_to_fastcs_datatype, snake_to_pascal
 
 
@@ -66,6 +66,7 @@ def test_numpy_to_fastcs_datatype(numpy_type, fastcs_datatype):
 
 def test_hinted_attributes_verified():
     loop = asyncio.get_event_loop()
+    transport_options = []
 
     class ControllerWithWrongType(Controller):
         hinted_wrong_type: AttrR[int]
@@ -75,7 +76,7 @@ def test_hinted_attributes_verified():
             self.attributes["hinted_wrong_type"] = self.hinted_wrong_type
 
     with pytest.raises(RuntimeError) as excinfo:
-        Backend(ControllerWithWrongType(), loop)
+        FastCS(ControllerWithWrongType(), transport_options, loop)
     assert str(excinfo.value) == (
         "Controller 'ControllerWithWrongType' introspection of hinted attribute "
         "'hinted_wrong_type' does not match defined datatype. "
@@ -86,7 +87,7 @@ def test_hinted_attributes_verified():
         hinted_int_missing: AttrR[int]
 
     with pytest.raises(RuntimeError) as excinfo:
-        Backend(ControllerWithMissingAttr(), loop)
+        FastCS(ControllerWithMissingAttr(), transport_options, loop)
     assert str(excinfo.value) == (
         "Controller `ControllerWithMissingAttr` failed to introspect hinted attribute "
         "`hinted_int_missing` during initialisation"
@@ -100,7 +101,7 @@ def test_hinted_attributes_verified():
             self.attributes["hinted"] = self.hinted
 
     with pytest.raises(RuntimeError) as excinfo:
-        Backend(ControllerAttrWrongAccessMode(), loop)
+        FastCS(ControllerAttrWrongAccessMode(), transport_options, loop)
     assert str(excinfo.value) == (
         "Controller 'ControllerAttrWrongAccessMode' introspection of hinted attribute "
         "'hinted' does not match defined access mode. Expected 'AttrR', got 'AttrRW'."
@@ -118,7 +119,7 @@ def test_hinted_attributes_verified():
         hinted_enum: AttrRW[MyEnum] = AttrRW(Enum(MyEnum2))
 
     with pytest.raises(RuntimeError) as excinfo:
-        Backend(ControllerWrongEnumClass(), loop)
+        FastCS(ControllerWrongEnumClass(), transport_options, loop)
     assert str(excinfo.value) == (
         "Controller 'ControllerWrongEnumClass' introspection of hinted attribute "
         "'hinted_enum' does not match defined datatype. "
